requirements.txt
Flask==2.3.2
google-generativeai==0.0.5
python-dotenv==1.0.0
PyPDF2==3.0.0
Werkzeug==2.3.4
